code refactor
Reorganized code into classes. AndSearch is functional and fully integrated/tested with docx, pdf, epub. Found txt compatibility problem.

GitHub is not tying together all of the relevant before-and-after lines of code. For example, it thinks I deleted Zayn's "class Search", but it just got pushed down a few lines, but is otherwise untouched. If you have questions I can help address them..
The goal of this change is to make it more modular and readable..